Updated `HumanPlayer` logic
- Enhanced `HumanPlayer` to validate promotion moves and fixed infinite loop logic.
- Included validation in `MakeMove` method to ensure moves are legal in debug mode.

+++ src/ChessBot.Core/Enums/PromotionFlag.cs
@@ -1,7 +1,7 @@
 ﻿namespace ChessBot.Core.Enums;
 
 [Flags]
-public enum PromotionFlag : uint
+public enum PromotionFlag : byte
 {
     None,
     Bishop = 2,
